# Backend/static/swagger_config.json
{
    "swagger": "2.0",
    "info": {
        "title": "PlayStation 5 E-commerce API",
        "description": "API documentation for the PlayStation 5 e-commerce website",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": ["http"],
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Enter 'Bearer' [space] and then your token in the text input below.\n\nExample: 'Bearer abcdef12345'"
        }
    },
    "paths": {
...
        "/products/categories": {
            "get": {
                "tags": ["Products"],
                "summary": "Get all product categories",
                "description": "Retrieves all product categories available.",
                "responses": {
                    "200": {
                        "description": "Categories retrieved successfully",
                        "schema": {
                            "type": "array",
                            "items": { "type": "string" }
                        }
                    },
                    "404": {
                        "description": "No categories found",
                        "schema": { "type": "string" }
                    }
                }
            }
        },
        "/products/category": {
            "post": {
                "tags": ["Products"],
                "summary": "Create a new product category",
                "description": "Endpoint to create a new product category with authentication and permission checks.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Category object that needs to be created",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": { "type": "string" }
                            },
                            "example": {
                                "name": "Newcategoryname"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category created successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": { "type": "string" }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input data",
                        "schema": {
                            "type": "array",
                            "items": { "type": "string" }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Authentication credentials missing or invalid",
                        "schema": { "type": "string" }
                    },
                    "403": {
                        "description": "Forbidden - Insufficient permissions",
                        "schema": { "type": "string" }
                    },
                    "404": {
                        "description": "Not Found - Endpoint not found",
                        "schema": { "type": "string" }
                    }
                },
                "security": [{ "Bearer": [] }]
            }
        },
        "/category/{category_id}": {
            "put": {
                "tags": ["Products"],
                "summary": "Update a product category",
                "description": "Endpoint to update a product category with authentication and permission checks.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "category_id",
                        "type": "integer",
                        "required": true,
                        "description": "Category ID"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Category object that needs to be updated",
                        "required": false,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": { "type": "string" },
                                "id": { "type": "integer" }
                            },
                            "example": {
                                "name": "UpdatedCategoryName",
                                "id": 1
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category updated successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": { "type": "string" },
                                "id": { "type": "integer" }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input data",
                        "schema": {
                            "type": "array",
                            "items": { "type": "string" }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Authentication credentials missing or invalid",
                        "schema": { "type": "string" }
                    },
                    "403": {
                        "description": "Forbidden - Insufficient permissions",
                        "schema": { "type": "string" }
                    },
                    "404": {
                        "description": "Not Found - Category not found",
                        "schema": { "type": "string" }
                    }
                },
                "security": [{ "Bearer": [] }]
            },
            "delete": {
                "tags": ["Products"],
                "summary": "Delete a product category",
                "description": "Endpoint to delete a product category with authentication and permission checks.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "category_id",
                        "type": "integer",
                        "required": true,
                        "description": "Category ID"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Category deleted successfully",
                        "schema": { "type": "string" }
                    },
                    "401": {
                        "description": "Unauthorized - Authentication credentials missing or invalid",
                        "schema": { "type": "string" }
                    },
                    "403": {
                        "description": "Forbidden - Insufficient permissions",
                        "schema": { "type": "string" }
                    },
                    "404": {
                        "description": "Not Found - Category not found",
                        "schema": { "type": "string" }
                    }
                },
                "security": [{ "Bearer": [] }]
            }
        }
    }
}